+++ Tests/UIKitTests/UIViewExtensionsTests.swift
@@ -252,15 +252,15 @@ final class UIViewExtensionsTests: XCTestCase {
         waitForExpectations(timeout: 0.5, handler: nil)
     }
 
-    func testRotateByAngle() {
+    func testRotateByRadians() {
         let view1 = UIView()
         let transform1 = CGAffineTransform(rotationAngle: 2)
-        view1.rotate(byAngle: 2, ofType: .radians, animated: false, duration: 0, completion: nil)
+        view1.rotate(byRadian: 2, animated: false, duration: 0, completion: nil)
         XCTAssertEqual(view1.transform, transform1)
 
         let view2 = UIView()
-        let transform2 = CGAffineTransform(rotationAngle: .pi * 90.0 / 180.0)
-        view2.rotate(byAngle: 90, ofType: .degrees, animated: false, duration: 0, completion: nil)
+        let transform2 = CGAffineTransform(rotationAngle: CGFloat(90).degreesToRadians)
+        view2.rotate(byRadian: CGFloat(90).degreesToRadians, animated: false, duration: 0, completion: nil)
         XCTAssertEqual(view2.transform, transform2)
 
         let rotateExpectation = expectation(description: "view rotated")
@@ -268,22 +268,22 @@ final class UIViewExtensionsTests: XCTestCase {
         let view3 = UIView()
         let transform3 = CGAffineTransform(rotationAngle: 2)
 
-        view3.rotate(byAngle: 2, ofType: .radians, animated: true, duration: 0.5) { _ in
+        view3.rotate(byRadian: 2, animated: true, duration: 0.5) { _ in
             rotateExpectation.fulfill()
         }
         XCTAssertEqual(view3.transform, transform3)
         waitForExpectations(timeout: 0.5, handler: nil)
     }
 
-    func testRotateToAngle() {
+    func testRotateByDegrees() {
         let view1 = UIView()
         let transform1 = CGAffineTransform(rotationAngle: 2)
-        view1.rotate(toAngle: 2, ofType: .radians, animated: false, duration: 0, completion: nil)
+        view1.rotate(byDegree: CGFloat(2).radiansToDegrees, animated: false, duration: 0, completion: nil)
         XCTAssertEqual(view1.transform, transform1)
 
         let view2 = UIView()
-        let transform2 = CGAffineTransform(rotationAngle: .pi * 90.0 / 180.0)
-        view2.rotate(toAngle: 90, ofType: .degrees, animated: false, duration: 0, completion: nil)
+        let transform2 = CGAffineTransform(rotationAngle: CGFloat(90).degreesToRadians)
+        view2.rotate(byDegree: 90, animated: false, duration: 0, completion: nil)
         XCTAssertEqual(view2.transform, transform2)
 
         let rotateExpectation = expectation(description: "view rotated")
@@ -291,7 +291,53 @@ final class UIViewExtensionsTests: XCTestCase {
         let view3 = UIView()
         let transform3 = CGAffineTransform(rotationAngle: 2)
 
-        view3.rotate(toAngle: 2, ofType: .radians, animated: true, duration: 0.5) { _ in
+        view3.rotate(byDegree: CGFloat(2).radiansToDegrees, animated: true, duration: 0.5) { _ in
+            rotateExpectation.fulfill()
+        }
+        XCTAssertEqual(view3.transform, transform3)
+        waitForExpectations(timeout: 0.5, handler: nil)
+    }
+
+    func testRotateToDegree() {
+        let view1 = UIView()
+        let transform1 = CGAffineTransform(rotationAngle: 2)
+        view1.rotate(toDegree: CGFloat(2).radiansToDegrees, animated: false, duration: 0, completion: nil)
+        XCTAssertEqual(view1.transform, transform1)
+
+        let view2 = UIView()
+        let transform2 = CGAffineTransform(rotationAngle: CGFloat(90).degreesToRadians)
+        view2.rotate(toDegree: 90, animated: false, duration: 0, completion: nil)
+        XCTAssertEqual(view2.transform, transform2)
+
+        let rotateExpectation = expectation(description: "view rotated")
+
+        let view3 = UIView()
+        let transform3 = CGAffineTransform(rotationAngle: 2)
+
+        view3.rotate(toDegree: CGFloat(2).radiansToDegrees, animated: true, duration: 0.5) { _ in
+            rotateExpectation.fulfill()
+        }
+        XCTAssertEqual(view3.transform, transform3)
+        waitForExpectations(timeout: 0.5, handler: nil)
+    }
+
+    func testRotateToRadian() {
+        let view1 = UIView()
+        let transform1 = CGAffineTransform(rotationAngle: 2)
+        view1.rotate(toRadian: 2, animated: false, duration: 0, completion: nil)
+        XCTAssertEqual(view1.transform, transform1)
+
+        let view2 = UIView()
+        let transform2 = CGAffineTransform(rotationAngle: CGFloat(90).degreesToRadians)
+        view2.rotate(toRadian: CGFloat(90).degreesToRadians, animated: false, duration: 0, completion: nil)
+        XCTAssertEqual(view2.transform, transform2)
+
+        let rotateExpectation = expectation(description: "view rotated")
+
+        let view3 = UIView()
+        let transform3 = CGAffineTransform(rotationAngle: 2)
+
+        view3.rotate(toRadian: 2, animated: true, duration: 0.5) { _ in
             rotateExpectation.fulfill()
         }
         XCTAssertEqual(view3.transform, transform3)
@@ -312,6 +358,22 @@ final class UIViewExtensionsTests: XCTestCase {
 
         XCTAssertEqual(view1.transform, view2.transform)
         XCTAssertEqual(view1.transform, view3.transform)
+    }
+
+    func testHorizontalShake() {
+        let view = UIView()
+        view.horizontalShake(animationType: .easeIn)
+        let currentAnimation = view.layer.animation(forKey: view.layer.animationKeys()?.first ?? "")
+
+        XCTAssertEqual(currentAnimation?.timingFunction, CAMediaTimingFunction(name: .easeIn))
+    }
+
+    func testVerticalShake() {
+        let view = UIView()
+        view.verticalShake(animationType: .linear)
+        let currentAnimation = view.layer.animation(forKey: view.layer.animationKeys()?.first ?? "")
+
+        XCTAssertEqual(currentAnimation?.timingFunction, CAMediaTimingFunction(name: .linear))
     }
 
     func testRemoveSubviews() {